Update
- Edit translate files.
  - Added new translation keys.
- Added tags: `amulet_of_creation_nether_stage_biomes`, `amulet_of_creation_end_stage_biomes`, `amulet_of_creation_overworld_stage_biomes`, `amulet_of_creation_final_stage_biomes` and `amulet_of_creation_final_stage_dimensions` for change the places where you can improve the amulet of creation.
- Yhe amulet of creation is complete.
- Deleting `BWModelPredicateProviders`.
- Now, if an entity comes in contact with an infected unit, it gets the effect of being infected.

+++ src/main/java/com/skylightmodding/items/components/ToolModifications.java
@@ -74,7 +74,7 @@ public class ToolModifications {
                 isSmelting = true;
                 serverWorld.spawnParticles(ParticleTypes.FLAME, (double) pos.getX() + 0.5D, (double) pos.getY() + 0.5D, (double) pos.getZ() + 0.5D, 20, 0.15D, 0.25D, 0.15D, 0.035f);
                 serverWorld.playSound(null, (double) pos.getX() + 0.5D, (double) pos.getY() + 0.5D, (double) pos.getZ() + 0.5D, SoundEvents.BLOCK_FIRE_EXTINGUISH, SoundCategory.BLOCKS, 0.4F, 2.6F);
-                for (int i = 0; i < inputList.getFirst().getCount(); i++) {
+                for (byte i = 0; i < inputList.getFirst().getCount(); i++) {
                     Block.dropStack(world, pos, result);
                 }
             }
@@ -87,7 +87,7 @@ public class ToolModifications {
 
 
     /** Поджигание атакуемых мобов */
-    public static void torchTheEnemy(int seconds, LivingEntity enemy) {
+    public static void torchTheEnemy(byte seconds, LivingEntity enemy) {
         if (!enemy.isFireImmune()) {
             enemy.setOnFireFor(seconds);
         }
